+++ src/main/java/io/github/jass2125/sistema/alocacao/core/actions/holiday/LoadHolidayEditCommand.java
@@ -33,7 +33,7 @@ public class LoadHolidayEditCommand implements Command  {
     public String execute(HttpServletRequest request, HttpServletResponse response) {
         try {
             String id = request.getParameter("idHoliday");
-            int idHoliday = Integer.parseInt(id);
+            Long idHoliday = Long.parseLong(id);
             Factory factory = new FactoryDao();
             HolidayDao dao = factory.createHolidayDao();
             Holiday holiday = dao.findById(idHoliday);
